Fix issue of `ask` API.
